pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "scarab"
version = "0.1.0"
description = "Interactive workout sequencer TUI with ASCII animations"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "textual>=0.80.0",
    "textual-autocomplete>=4.0.0",
    "ruamel.yaml>=0.18.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]
ascii-gen = [
    "ascii-magic>=2.0.0",
    "pillow>=10.0.0",
]

[project.scripts]
scarab = "scarab.app:main"

[tool.hatch.build.targets.wheel]
packages = ["src/scarab"]

[tool.hatch.build.targets.sdist]
include = ["src/scarab", "README.md"]
